requirements.txt
python-telegram-bot
requests
logging